# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GestiCom API",
        "description": "Documentation de l'API GestiCom pour la gestion de caisse",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Serveur local"
        },
        {
            "url": "https://gestion-caisse.onrender.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/ventes": {
            "post": {
                "tags": [
                    "Ventes"
                ],
                "summary": "Créer une vente",
                "operationId": "createVente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lignes",
                                    "modePaiement"
                                ],
                                "properties": {
                                    "lignes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "produitId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantite": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "modePaiement": {
                                        "type": "string",
                                        "enum": [
                                            "especes",
                                            "carte",
                                            "mobileMoney"
                                        ],
                                        "example": "especes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vente créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vente effectuée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "reference": {
                                                    "type": "string",
                                                    "example": "V-0001"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": 3000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ventes/{id}/annuler": {
            "post": {
                "tags": [
                    "Ventes"
                ],
                "summary": "Annuler une vente",
                "operationId": "annulerVente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vente annulée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Vente annulée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}